+++ src/emails/emails.enums.ts
@@ -3,3 +3,9 @@ export enum EmailSearchSortOrder {
     ASC = 'asc',
     DESC = 'desc',
 }
+
+export enum EMAIL_STATUS {
+    RECEIVED = 'received',
+    DRAFT = 'draft',
+    SENT = 'sent',
+}